api-django/requirements.txt
Django>=4.2,<6
djangorestframework>=3.14
django-cors-headers>=4.3
python-dotenv>=1.0
